+++ src/logger.h
@@ -9,7 +9,7 @@ class Logger
 private:
 public:
     char logBuffer[LOG_BUFFER_SIZE];
-    uint32_t bytesLogged = 0;
+    unsigned long bytesLogged = 0;
 
     Logger();
     void log(const char *msg);
